Add an experimental implementation of k-means for cluster assignment.
This algothihm serves as an example of how to split vehicles into n sets of pods, inc such a way that each pod has a collection of vehicles that are collocated, and likely to interact.
Note that this has not been incorporated into the codebase that runs during normal Roadrunner operation.